+++ trikot-http/http/src/androidMain/kotlin/com/mirego/trikot/http/android/requestFactory/KtorHttpRequestFactory.kt
@@ -6,6 +6,7 @@ import com.mirego.trikot.http.HttpRequest
 import com.mirego.trikot.http.HttpRequestFactory
 import com.mirego.trikot.http.HttpResponse
 import com.mirego.trikot.http.RequestBuilder
+import com.mirego.trikot.http.exception.HttpNetworkUnreachableException
 import com.mirego.trikot.http.exception.HttpRequestTimeoutException
 import com.mirego.trikot.streams.cancellable.CancellableManager
 import com.mirego.trikot.streams.reactive.Publishers
@@ -31,6 +32,9 @@ import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import org.reactivestreams.Publisher
+import java.net.ConnectException
+import java.net.NoRouteToHostException
+import java.net.UnknownHostException
 import kotlin.coroutines.CoroutineContext
 import kotlin.time.Duration
 import kotlin.time.Duration.Companion.seconds
@@ -119,6 +123,9 @@ class KtorHttpRequestFactory(
                     } else {
                         publisher.error = when (ex) {
                             is KtorRequestTimeoutException -> HttpRequestTimeoutException(ex)
+                            is UnknownHostException,
+                            is ConnectException,
+                            is NoRouteToHostException -> HttpNetworkUnreachableException(ex)
                             else -> ex
                         }
                     }