+++ trikot-http/http/src/commonMain/kotlin/com/mirego/trikot/http/exception/HttpNetworkUnreachableException.kt
@@ -1,0 +1,3 @@
+package com.mirego.trikot.http.exception
+
+class HttpNetworkUnreachableException(source: Throwable) : Exception(source)