+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Updates
 
+- [http] Handle network resolution failures with new `HttpNetworkUnreachableException`
+
 ### Breaking Changes
 
 ## 6.0.0